login link on home Election param supported (#28)
This patch adds the support for the election presentation key
"show_login_link_on_home". In the csv file you should use the key
"Login link on home".

I've also added a new helper function to search the key ignoring the
upper or lower case, so it's possible to define this new option as
"LOGIN LINK ON HOME", "login link on home" or any other upper/lower case
combination.